+++ spec/functional/apns_spec.rb
@@ -107,6 +107,10 @@ describe 'APNs' do
         on.notification_id_will_retry do |_, n_id|
           retry_ids << n_id
         end
+
+        on.notification_will_retry do |n|
+          retry_ids << n.id
+        end
       end
 
       Rpush.embed
@@ -124,6 +128,18 @@ describe 'APNs' do
       wait_for_notification_to_deliver(notification)
       fail_notification(notification)
       wait_for_notification_to_fail(notification)
+    end
+
+    describe 'with a failed connection' do
+      it 'retries all notifications' do
+        Rpush::Daemon::TcpConnection.any_instance.stub(sleep: nil)
+        expect(ssl_socket).to receive(:write).exactly(4).times.and_raise(Errno::EPIPE)
+        notifications = nil
+        ActiveRecord::Base.transaction do
+          notifications = 2.times.map { create_notification }
+        end
+        notifications.each { |n| wait_for_notification_to_retry(n) }
+      end
     end
 
     describe 'with multiple notifications' do